requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9

python-dotenv~=0.20.0
aiogram~=2.20
gino~=1.0.1
SQLAlchemy~=1.3.24
geopy~=2.2.0